requirements.txt
flask
openai
requests
python-dotenv
PyPDF2